+++ src/ui/Features/Video/ReEncodeVideo/ReEncodeVideoViewModel.cs
@@ -324,7 +324,7 @@ public partial class ReEncodeVideoViewModel : ObservableObject
         if (long.TryParse(arr[1].Trim(), out var f))
         {
             _processedFrames = f;
-            ProgressValue = (double)_processedFrames * 100.0 / JobItems[_jobItemIndex].TotalFrames;
+            ProgressValue = _processedFrames * 100.0 / JobItems[_jobItemIndex].TotalFrames;
         }
     }
